Rework memory range used in bootloader
Not convinced the forced allocation from 0x20000 to 0x28000 for the heap
pages is safe but we'll roll with it for now. Since we map that range
for the kernel to use and it's within the first 4MiB, we don't want
other data to be saved there since it'll conflict with the identity
mapping. As such, for most allocations, use MaxAddress with a max of
0x20000