Adjuste memory allocation logic in bootloader
Previously, we would only load the kernel image to the first 0x18000
bytes. This heavily limits the maximum size of the kernel image, even
ignore memory contention from other data structures.

The new logic now:
1) Loads phdrs with a vaddr in the first 4MiB to the first 4MiB
(identity mapping will map those pages)
2) Loads other phdrs *anywhere*. This removes the size restriction on
the kernel image.
3) The stack/heap pages can now be allocated anywhere and will be mapped
to the correct virtual address.

Note: There is still a weird behaviour where, if something was to be
allocated in the 0x80000 range, the stack page would overwrite the
identity mapping, making the data in that address range unreadable. An
easy fix would be to move the virtual address of the stack out of the
4MiB that are identity mapped (to do later)